OSX, C, refact: move mem functions in mem.c
